pyproject.toml
[tool.poetry]
name = "realtime_jcdecaux_pipeline"
version = "0.1.0"
description = "Real-time data pipeline for JCDecaux bike stations"
authors = ["Yassine"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^2.4.2"
requests = "^2.31.0"
kafka-python = "^2.0.2"
elasticsearch = "^8.8.2"
pyspark = "^3.2.4"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C"]
